refactor(frontend/components): make refactoring for component files
