+++ frontend/src/components/AccountWrap/AccountWrap.scss
@@ -29,8 +29,6 @@
         }
     }
 
-    &__icon {}
-
     &__content {
         padding: 10px;
     }
