+++ frontend/src/components/FormWrap/FormWrap.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { useState } from "react"
-import FormRegistration from "../FormRegistration/FormRegistration"
-import FormAuthorization from "../FormAuthorization/FormAuthorization"
+import FormRegistration from "../Forms/FormRegistration/FormRegistration"
+import FormAuthorization from "../Forms/FormAuthorization/FormAuthorization"
 import signupIcon from "../../images/icons/signup.png"
 import loginIcon from "../../images/icons/login.png"
 import "./FormWrap.scss"
