+++ frontend/src/components/Forms/Form.scss
@@ -7,8 +7,6 @@
     padding: 12px;
     row-gap: 20px;
 
-    &__autorization {}
-
     &__label {
         display: flex;
         position: relative;
