+++ frontend/src/components/Forms/FormRegistration/FormRegistration.jsx
@@ -1,10 +1,10 @@
-import React, { useEffect } from "react"
+import React from "react"
 import { useForm } from "react-hook-form"
 import { yupResolver } from "@hookform/resolvers/yup"
 import * as yup from "yup"
 import axios from "axios"
 import { toast } from 'react-toastify'
-import "./FormRegistration.scss"
+import "../Form.scss"
 
 export default function FormRegistration({ changeMode }) {
     const schema = yup.object().shape({
